minor modification

+++ components/decoding.py
@@ -137,7 +137,7 @@ def beam_search_decode(get_logits_fn, init_cache, init_seq, init_seq_len, beam_s
                     # [Li+ 2016] "A simple, fast diverse decoding algorithm" with
                     # a fixed diversity rate.
                     top_logits, ids = tf.math.top_k(logits, beam_size, False, name='pre_tops') 
-                    diversify_bias = tf.range(beam_size) * params["diversity_rate"]
+                    diversify_bias = tf.cast(tf.range(beam_size), tf.float32) * params["diversity_rate"]
                     top_logits -= diversify_bias[None]
                 else:
                     assert False
